refactor, organize dialogs
