+++ Gruntfile.js
@@ -23,7 +23,7 @@ module.exports = function (grunt) {
   'src/ui/dialogs/scoreDialogs.js','src/ui/help.js',
    'src/ui/dialogs/fontComponent.js', 'src/ui/dialogs/staffComponents.js','src/ui/dialogs/textComponents.js',
    'src/ui/dialogs/textDialogs.js',
-   'src/ui/dialogs/measureDialogs.js','src/ui/dialogs/staffDialogs.js','src/ui/dialogs/libraryDialog.js','src/ui/dialogs/treeComponent.js' ,
+   'src/ui/dialogs/staffDialogs.js','src/ui/dialogs/libraryDialog.js','src/ui/dialogs/treeComponent.js' ,
    'src/ui/dialogs/fileDialogs.js',
    'src/ui/application.ts', 'src/ui/fileio/fileInput.js', 'src/ui/fileio/xhrLoader.js','src/ui/fileio/library.ts',
   ]
