+++ src/ui/i18n/language_en.json
@@ -905,7 +905,7 @@
       }
      },
      {
-      "ctor": "SuiTextTransformDialog",
+      "ctor": "SuiTextBlockDialog",
       "label": "Text Properties",
       "dialogElements": [
        {
